requirements.txt
streamlit==1.0.0
numpy==1.18.5
pickleshare==0.7.5 
sklearn==0.0
Pillow==8.4.0
